# Experiment: speed up decideFin! by optimizing terms