--- a/equational_theories/RSimpSet.lean
+++ b/equational_theories/RSimpSet.lean
@@ -1,0 +1,4 @@
+import Lean
+
+-- For DecideBang, but has to be in its own file
+register_simp_attr rsimp